requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.17.1
click==8.1.8
contextlib2==21.6.0
cryptography==44.0.2
Django==5.2
django-cors-headers==4.7.0
django-ninja==1.4.1
django-ninja-extra==0.30.0
django-ninja-jwt==5.3.7
dnspython==2.7.0
email_validator==2.2.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
injector==0.22.0
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
